Refactor: useInfiniteScroll 커스텀함수로 추상화
